+++ packages/designer/src/builtin-simulator/renderer.ts
@@ -1,9 +1,9 @@
 import { ComponentInstance, NodeInstance, Component } from '../simulator';
-import { ComponentSchema } from '@ali/lowcode-types';
+import { NodeSchema } from '@ali/lowcode-types';
 
 export interface BuiltinSimulatorRenderer {
   readonly isSimulatorRenderer: true;
-  createComponent(schema: ComponentSchema): Component | null;
+  createComponent(schema: NodeSchema): Component | null;
   getComponent(componentName: string): Component;
   getComponentInstances(id: string): ComponentInstance[] | null;
   getClosestNodeInstance(from: ComponentInstance, nodeId?: string): NodeInstance<ComponentInstance> | null;